Support for appending extra GCC attributes to symbols.
For a reason I have not yet diagnosed, GCC attributes appear mangled in the
Clang AST. Specifically, when we're walking it, we see the '__attribute__'
token, but not the contents of the attribute itself. In the code of the pruner
we just strip this stuff to avoid emitting malformed C. For the most part this
has no adverse effects. However, some attributes are necessary for correctness.
Specifically, CAmkES dataports carry a specific alignment that is necessary for
remapping the underlying frames appropriately.

This commit adds support for the user to re-add attributes to symbols. No real
validation is done here and it is expected that the user only adds sensible
attributes to sensible entities. If you induce malformed C or inappropriate
attributes using this mechanism, it is your own fault.